Add Descr. and Ex. for RNN variants
RNN.ipynb is a notebook that describes the main properties of vanilla RNNs, as well as GRU- and LSTM-based RNNs. For an annotated example of how to use GRU-based NNs in TensorFlow for natural language processing, run GRU_Example_Code.py with train.txt as the third argument and dev.txt as the fourth argument in your the command line (python GRU_Example_Code.py train.txt dev.txt).